refactor: To move authentication operation from crud to auth structure.
