+++ samudra/core/crud/__init__.py
@@ -2,4 +2,3 @@ from .lemma import *
 from .konsep import *
 from .cakupan import *
 from .kata_asing import *
-from .user import *
